fix type

+++ pkg/domain/merakiWebhook.go
@@ -28,8 +28,8 @@ type TriggerData struct {
 }
 
 type Trigger struct {
-	Ts          int    `json:"ts"`
-	Type        string `json:"type"`
-	NodeId      int    `json:"nodeId"`
-	SensorValue int    `json:"sensorValue"`
+	Ts          int     `json:"ts"`
+	Type        string  `json:"type"`
+	NodeId      int     `json:"nodeId"`
+	SensorValue float64 `json:"sensorValue"`
 }
